ECampus/src/ECampus.Infrastructure/DataSelectors/MultipleItemSelectors/MultipleDepartmentSelector.cs
using ECampus.Infrastructure.Interfaces;
using ECampus.Shared.Extensions;
using ECampus.Shared.Models;
using ECampus.Shared.QueryParameters;
using Microsoft.EntityFrameworkCore;

namespace ECampus.Infrastructure.DataSelectors.MultipleItemSelectors;

public class MultipleDepartmentSelector : IMultipleItemSelector<Department, DepartmentParameters>
{
    public IQueryable<Department> SelectData(ApplicationDbContext context, DepartmentParameters parameters)
    {
        var result = context.Departments.Search(d => d.Name, parameters.DepartmentName);
        if (parameters.FacultyId == 0)
        {
            return result;
        }
        return result.Where(d => d.FacultyId == parameters.FacultyId);
    }
}